Move `RunModuleManager` to its own file.
