+++ packages/jsandbox/src/index.ts
@@ -4,47 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-  ConsoleStdout,
-  OpenFile,
-  WASI,
-  File as WasiFile,
-} from "@bjorn3/browser_wasi_shim";
-import { type InputValues, type Kit } from "@google-labs/breadboard";
-
-import factory from "./factory.js";
-
-export { factory, RunModuleManager };
-
-class RunModuleManager {
-  constructor(public readonly runtimeUrl: URL) {}
-
-  async runModule(code: string, inputs: InputValues) {
-    const wasi = new WASI(
-      [],
-      [],
-      [
-        new OpenFile(new WasiFile([])), // stdin
-        ConsoleStdout.lineBuffered((msg) =>
-          console.log(`[WASI stdout] ${msg}`)
-        ),
-        ConsoleStdout.lineBuffered((msg) =>
-          console.warn(`[WASI stderr] ${msg}`)
-        ),
-      ]
-    );
-    const jsandbox = factory();
-    const { instance } = await WebAssembly.instantiateStreaming(
-      fetch(this.runtimeUrl),
-      {
-        "./jsandbox_bg.js": jsandbox,
-        wasi_snapshot_preview1: wasi.wasiImport,
-      }
-    );
-    jsandbox.__wbg_set_wasm(instance.exports);
-    // @ts-expect-error 2739
-    wasi.start({ exports: instance.exports });
-    const result = jsandbox.run_module(code, JSON.stringify(inputs));
-    return JSON.parse(result);
-  }
-}
+export { RunModuleManager } from "./browser.js";
